+++ include/game/data.h
@@ -33,32 +33,41 @@ struct GameData {
     };
 
     std::unordered_map<int, EntityData> entitiesData = { // health, attack, defence, speed, crit, pickup, regeneration
-        { 0, { "player.png", { 10.0f, 1.0f, 0.0f, 0.05f, 5.0f, 2.0f, 0.0f }, H2DE_LevelRect{ 0.0f, 0.0f, 1.0f, 1.0f }, H2DE_LevelRect{ 0.25f, 0.25f, 0.5f, 0.5f }, 0}},
-        { 1, { "player.png", { 15.0f, 1.0f, 0.0f, 0.02f, 5.0f, 1.0f, 0.0f }, std::nullopt, H2DE_LevelRect{ 0.25f, 0.25f, 0.5f, 0.5f }, 1}},
+        { 0, { "player.png", { 1.0f, 1.0f }, { 100, 1, 0, 0.05f, 5.0f, 2.0f, 1 }, H2DE_LevelRect{ 0.0f, 0.0f, 1.0f, 1.0f }, H2DE_LevelRect{ 0.25f, 0.25f, 0.5f, 0.5f }, 0}},
+        { 1, { "player.png", { 1.0f, 1.0f }, { 15, 1, 0, 0.02f, 5.0f, 1.0f, 0 }, std::nullopt, H2DE_LevelRect{ 0.25f, 0.25f, 0.5f, 0.5f }, 1}},
     };
 
     std::unordered_map<int, WeaponData> weaponsData = { // speed, damage, piercing, explosive
         { 0, { "9mm", "weapon.png", 500, {
             "9mm-bullet.png", { 0.0f, 0.0f, 0.15f, 0.15f },
-                0.15f, 10.0f, false, false
+                0.15f, 10, false, false
             }
         }},
         { 1, { "m4", "weapon.png", 200, {
             "9mm-bullet.png", { 0.0f, 0.0f, 0.15f, 0.15f },
-                0.2f, 8.0f, false, false
+                0.2f, 8, false, false
             }
         }},
     };
 
+    H2DE_LevelSize healthBarSize = { 6.0f, 0.4f };
+    H2DE_LevelSize xpBarSize = { 6.0f, 0.4f };
+
     float nearPlayer = 0.2f;
     float critDamageMultiplier = 1.5f;
     float itemAirResistance = 0.005f;
-    float itemSpeed = 0.1f;
+    float itemSpeed = 0.07f;
     float enemyDropingCoin = 5.0f;
     float pickUpItemSpeedMultiplier = 1.05f;
+    float cameraPadding = 0.4f;
+    float innerPadding = 0.2f;
+    float levelGrowthFactor = 1.15f;
+    float bigEnemySizeMultiplier = 1.2f;
+    float bigEnemyStatsMultiplier = 3.0f;
 
     int enemyAttackDelay = 600;
     int maxXpLevel = 6;
+    int firstLevelNeededXp = 10;
 };
 
 #endif