+++ include/game/game.h
@@ -16,8 +16,10 @@ private:
     std::vector<SDL_Keycode> keysDown = {};
     std::vector<int> mouseButtonsDown = {};
     H2DE_AbsPos mousePos = { 0, 0 };
+    bool debug = false;
 
     void initEngine();
+    void initFont();
     void initMap();
     void initCamera();
 
@@ -36,6 +38,7 @@ public:
     std::vector<SDL_Keycode> getKeysDown() const;
     std::vector<int> getMouseButtonsDown() const;
     H2DE_AbsPos getMousePos() const;
+    bool isDebuging() const;
 };
 
 #endif