+++ include/game/map.h
@@ -1,10 +1,14 @@
 #ifndef MAP_H
 #define MAP_H
 
+#include <iomanip>
+#include <sstream>
 #include "game.h"
+#include "interface.h"
 #include "entity.h"
 #include "items.h"
 class Game;
+class Interface;
 class Player;
 class Enemy;
 class Item;
@@ -13,6 +17,7 @@ class Bullet;
 class Map {
 private:
     Game* game;
+    Interface* ui;
     
     Player* player;
     std::vector<Enemy*> enemies;
@@ -22,7 +27,10 @@ private:
     H2DE_Timeline* t = nullptr;
     std::vector<H2DE_Timeline*> damageTimelines;
 
+    int killCount = 0;
+
     void initPlayer();
+    void initInterface();
     void generate();
     H2DE_TextureData getGroundTileTextureData(TileData tileData, TextureData textureData) const;
     H2DE_TextureData getDecorationTileTextureData(TileData tileData, TextureData textureData) const;
@@ -32,16 +40,21 @@ private:
     void destroyTiles();
     void destroyEnemies();
     void destroyPlayer();
+    void destroyDamageTimelines();
+    void destroyInterface();
 
     void updatePlayer();
     void updateEnemies();
     void updateItems();
     void updateDamageDisplay();
+    void updateUi();
 
-    void summonEnemy(int id, float size);
-    void summonEnemy(int id, H2DE_LevelPos pos, float size);
+    void summonEnemy(int id, bool big);
+    void summonEnemy(int id, H2DE_LevelPos pos, bool big);
 
     std::vector<Bullet*> getAllBullets() const;
+    int xpForLevel(unsigned int level) const;
+    int totalXpForLevel(unsigned int level) const;
 
 public:
     Map(Game* game);
@@ -51,14 +64,17 @@ public:
 
     void dropXp(H2DE_LevelPos pos, int level);
     void dropCoin(H2DE_LevelPos pos);
+    void enemyKilled();
 
-    void displayDamages(H2DE_LevelPos pos, float damages, H2DE_ColorRGB color);
+    void displayHealthActivity(H2DE_LevelPos pos, int healthActivity, H2DE_ColorRGB color);
 
     static int getIndex(float yPos, int index);
+    Interface* getUi() const;
     Player* getPlayer() const;
     Enemy* getEnemy(H2DE_LevelObject* object) const;
     Bullet* getBullet(H2DE_LevelObject* object) const;
     std::vector<Item*> getItems() const;
+    float getXpLevelPercentage() const;
 };
 
 #endif