+++ src/entities/enemy.cpp
@@ -19,14 +19,18 @@ Enemy::Enemy(Game* g, Map* m, H2DE_LevelPos p, EntityData d) : Entity(g, m, p, d
 
 // CLEANUP
 Enemy::~Enemy() {
-    std::cout << "Enemy cleared" << std::endl;
+    if (game->isDebuging()) std::cout << "Enemy cleared" << std::endl;
 }
 
 // EVENTS
 bool Enemy::isNearPlayer() {
     static float nearPlayer = game->getData()->nearPlayer;
-    H2DE_LevelPos pos = getObjectData()->pos;
-    H2DE_LevelPos playerPos = map->getPlayer()->getObjectData()->pos;
+
+    H2DE_LevelObjectData* thisData = getObjectData();
+    H2DE_LevelObjectData* playerData = map->getPlayer()->getObjectData();
+
+    H2DE_LevelPos pos = thisData->pos + thisData->hitboxes["damage"].rect.getCenter();
+    H2DE_LevelPos playerPos = playerData->pos + playerData->hitboxes["damage"].rect.getCenter();
 
     return (std::abs(playerPos.x - pos.x) < nearPlayer && std::abs(playerPos.y - pos.y) < nearPlayer);
 }
@@ -48,17 +52,25 @@ void Enemy::killImpl() {
     static float enemyDropingCoinProbability = game->getData()->enemyDropingCoin;
     H2DE_LevelObjectData* objData = getObjectData();
 
-    map->dropXp(objData->pos, data.xpOnDeath);
+    map->dropXp(objData->pos + objData->hitboxes["damage"].rect.getCenter(), data.xpOnDeath);
     map->getPlayer()->increaseCoins(1);
     if (H2DE_RandomFloatInRange(0.0f, 100.0f) <= enemyDropingCoinProbability) map->dropCoin(objData->pos);
+
+    map->enemyKilled();
+}
+
+void Enemy::inflictDamagesImpl(int damages, bool isCrit) {
+    
 }
 
 // UPDATE
 void Enemy::updateImpl() {
     if (!isNearPlayer()) {
-        static H2DE_LevelObjectData* playerData = map->getPlayer()->getObjectData();
-        H2DE_LevelPos pos = getObjectData()->pos;
-        H2DE_LevelPos playerPos = playerData->pos;
+        H2DE_LevelObjectData* thisData = getObjectData();
+        H2DE_LevelObjectData* playerData = map->getPlayer()->getObjectData();
+
+        H2DE_LevelPos pos = thisData->pos + thisData->hitboxes["damage"].rect.getCenter();
+        H2DE_LevelPos playerPos = playerData->pos + playerData->hitboxes["damage"].rect.getCenter();
 
         float angle = std::atan2(playerPos.y - pos.y, playerPos.x - pos.x);
         velocity = { data.stats.speed * std::cos(angle), data.stats.speed * std::sin(angle) };